graph/flow: rename cfa package to flow
Based on the expressed wishes of @kortschak in https://github.com/gonum/gonum/pull/803#issuecomment-455102383
